# 2 pixel-aligned vertical lines shortcut